+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Changed
+
+- Change messages ids to be the same as the ones from Checkout API
+
 ## [0.4.4] - 2019-09-11
 
 ## Fixed